designs: add default designs

+++ packages/designs/src/deleteDesign/deleteDesign.ts
@@ -1,6 +1,8 @@
 import { Events } from '@minddrop/events';
 import { Fs } from '@minddrop/file-system';
+import { InvalidParameterError } from '@minddrop/utils';
 import { DesignsStore } from '../DesignsStore';
+import { defaultDesignIds } from '../default-designs';
 import { DesignDeletedEvent, DesignDeletedEventData } from '../events';
 import { getDesign } from '../getDesign';
 
@@ -12,6 +14,13 @@ import { getDesign } from '../getDesign';
 export async function deleteDesign(id: string): Promise<void> {
   // Get the design
   const design = getDesign(id);
+
+  // It prevents deleting default designs
+  if (defaultDesignIds.includes(design.id)) {
+    throw new InvalidParameterError(
+      `Cannot delete default design ${design.id}`,
+    );
+  }
 
   // Delete the design file
   await Fs.removeFile(design.path);
